Buy/sell with placeholders
Placeholders:
- user funds account balance
- current crypto value

+++ src/main/java/org/main/criptoapi/mvtCrypto/MtvCrypto.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.Setter;
 import org.main.criptoapi.crypto.Crypto;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -35,6 +34,6 @@ public class MtvCrypto {
     private LocalDateTime daty;
 
     @Column(name = "valeur", precision = 15, scale = 2)
-    private BigDecimal valeur;
+    private Double valeur;
 
 }
